fix(auth): change employeeId type from Integer to Long in UserPrincipal class

+++ Tick-It-Notification-Service/src/main/java/com/harish/tickit/notificationService/auth/UserPrincipal.java
@@ -2,13 +2,13 @@ package com.harish.tickit.notificationService.auth;
 
 public class UserPrincipal
 {
-	private Integer employeeId;
+	private Long employeeId;
 	private String userName;
 	
-	public Integer getEmployeeId() {
+	public Long getEmployeeId() {
 		return employeeId;
 	}
-	public void setEmployeeId(Integer employeeId) {
+	public void setEmployeeId(Long employeeId) {
 		this.employeeId = employeeId;
 	}
 	public String getUserName() {
